requirements.txt
asgiref==3.3.1
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.1
typing_extensions==4.8.0
